+++ CHANGELOG.md
@@ -75,6 +75,7 @@ General Public License v2 (GPLv2). See LICENSES directory or go to
 - Crash when use Btrfs-subvolume as backup destination (Daidalos [@D4id4los](https://github.com/D4id4los), [#2487](https://github.com/bit-team/backintime/issues/2487))
 - Crash when polkit authentication dialog times out while setting up udev scheduling ([#2375](https://github.com/bit-team/backintime/issues/2375))
 - Guard pause/resume/stop actions in the GUI against ProcessLookupError when the backup process has already exited. ([#1604](https://github.com/bit-team/backintime/issues/1604))
+* Enum names in config file are replaced by their values
 
 ## [1.6.1] (2026-02-10)